Boost: first pass of boost refactoring
