+++ js/modules/boost/boostables.js
@@ -1,0 +1,26 @@
+/**
+ *
+ * Copyright (c) 2019-2019 Highsoft AS
+ *
+ * Boost module: stripped-down renderer for higher performance
+ *
+ * License: highcharts.com/license
+ *
+ */
+
+// These are the series we allow boosting for.
+
+var boostables = [
+    'area',
+    'arearange',
+    'column',
+    'columnrange',
+    'bar',
+    'line',
+    'scatter',
+    'heatmap',
+    'bubble',
+    'treemap'
+];
+
+export default boostables;
